+++ src/blosc2/__init__.py
@@ -588,20 +588,25 @@ from .lazyexpr import (
     result_type,
     can_cast,
 )
-from .proxy import (
-    Proxy,
+from .proxy_source import (
     ProxySource,
     ProxyNDSource,
-    ProxyNDField,
     ByteRangeNDSource,
     FsspecNDSource,
-    SimpleProxy,
-    jit,
-    as_simpleproxy,
 )
 from .indexing import Index
 
 from .schunk import SChunk, load, open
+
+# After .schunk and .indexing, which it may now import at module level: what a
+# proxy reads from lives in .proxy_source, which the early modules reach instead
+from .proxy import (
+    Proxy,
+    ProxyNDField,
+    SimpleProxy,
+    jit,
+    as_simpleproxy,
+)
 from . import linalg
 from .linalg import tensordot, vecdot, permute_dims, matrix_transpose, matmul, transpose, diagonal, outer
 from .utils import linalg_funcs as linalg_funcs_list
