+++ tests/ndarray/test_proxy.py
@@ -375,3 +375,16 @@ def test_vlmeta_cannot_overwrite_proxy_state():
     # Anything else still goes through
     proxy = blosc2.Proxy(source, vlmeta={"mine": "ok"})
     assert proxy.vlmeta["mine"] == "ok"
+
+
+def test_the_proxy_module_still_answers_for_the_source_names():
+    # They live in `blosc2.proxy_source` now, so that the modules bound early in
+    # `blosc2/__init__` can reach them without dragging `proxy.py` in ahead of
+    # `schunk` and `indexing`.  `blosc2.proxy.X` is where anything outside would
+    # look for them, so that keeps working: the imports there are not dead.
+    import blosc2.proxy
+    import blosc2.proxy_source
+
+    for name in ("ProxySource", "ProxyNDSource", "ByteRangeNDSource", "FsspecNDSource"):
+        assert getattr(blosc2.proxy, name) is getattr(blosc2.proxy_source, name)
+        assert getattr(blosc2, name) is getattr(blosc2.proxy_source, name)
